+++ src/Service/Controllers/ArtistController.cs
@@ -9,6 +9,7 @@ using log4net;
 
 using LMS.Model;
 using LMS.Model.Resource;
+using LMS.Model.Composite;
 using LMS.Data;
 
 namespace LMS.Service.Controllers
@@ -97,6 +98,20 @@ namespace LMS.Service.Controllers
 
             try
             {
+                // If artist has any related resources, then artist cannot be deleted.
+                // Related resource; Account
+                IRepository<Account> accountRepository = RepositoryFactory<Account>.Create();
+                List<Account> accounts = accountRepository.Find(new Account() { Artist = new Reference(Reference.ArtistUri, id) });
+                if (accounts != null && accounts.Count > 0)
+                    return BadRequest("Artist cannot be deleted; one or more accounts are related to artist.");
+
+                // Related resource; Statement
+                IRepository<Statement> statementRepository = RepositoryFactory<Statement>.Create();
+                List<Statement> statements = statementRepository.Find(new Statement() { Artist = new Reference(Reference.ArtistUri, id) });
+                if (statements != null && statements.Count > 0)
+                    return BadRequest("Artist cannot be deleted; one or more statements are related to artist.");
+
+                // Delete artist resource.
                 IRepository<Artist> repository = RepositoryFactory<Artist>.Create();
                 repository.Remove(id);
             }
